pyproject.toml
[project]
name = "precificador-imobiliario"
version = "0.1.0"
description = "Sistema de precificação de terrenos urbanos em SP"
requires-python = ">=3.10"
dependencies = [
    "pandas==1.5.3",
    "numpy>=1.23.5,<2.0.0",
    "scikit-learn==1.2.2",
    "lightgbm==3.3.5",
    "xgboost==1.7.6",
    "fastapi==0.110.0",
    "uvicorn==0.29.0",
    "jinja2==3.1.3",
    "mlflow==2.12.1",
    "dvc==3.50.1",
    "beautifulsoup4==4.12.3",
    "openpyxl==3.1.2",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "evidently==0.4.33",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.72",
    "langchain-text-splitters==0.3.9",
    "langchain-google-genai==2.1.8",
    "langgraph==0.6.3",
    "chromadb==0.5.23",
    "sentence-transformers==3.3.1",
    "pypdf==5.1.0",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
